Add robust minimum shortfall objective

Solves a minimum shortfall objective that minimises a robust shortfall objective across all features. For each feature, the shortfall objective is quantified based on the distribution of shortfall across all realizations of data. For a fully robust solution, the shortfall metric used for optimization is the maximum shortfall across all realizations of the data. For a partially robust solution (conf_level < 1), the solution guarantees that the probability that the shortfall is greater than the shortfall metric used for optimisation is less than or equals to 1 - conf_level.

Usage

add_robust_min_shortfall_objective(x, budget)

Arguments

x

prioritizr::problem() object.

budget

numeric value specifying the maximum expenditure of the prioritization. For problems with multiple zones, the argument to budget can be (i) a single numeric value to specify a single budget for the entire solution or (ii) a numeric vector to specify a separate budget for each management zone.

method

the probabilistic constraint formulation method, either Chance (default) or CondValueAtRisk. See details.

Value

An updated prioritizr::problem() object with the objective added to it.

Details

Minimises an objective based on the total amount of shortfall (difference between the target and the feature representation) weighted across all features, subject to a budget.

For a feature with more than one number of realizations in the group, it computes the shortfall metric for that feature based on the following approach:

If conf_level = 1, the shortfall metric is the maximum of the shortfall across all data realizations in the group

If conf_level < 1, then the shortfall metric is:

Method = 'CondValueAtRisk': the average of the shortfall in the set of data realizations that exceed the 1-conf_level quantile

Method = 'Chance': the 1-conf_level quantile of the distribution of shortfall

Mathematical formulation

This objective can be expressed mathematically for a set of planning units (\(I\) indexed by \(i\)), a set of features (\(J\) indexed by \(j\)), and a set of realizations (\(K\) indexed by \(k\)) as:

$$\mathit{Minimize} \space \sum_{j = 1}^{J} w_j y_j \\ \mathit{subject \space to} \\ \Pr_k\{\sum_{i = 1}^{I} x_i r_{ijk} + T_{j} y_j \geq T_j\} > \alpha \qquad \forall j \in J \\ \sum_{i = 1}^{I} x_i c_i \leq B \\ y_j \geq v_{jk} \qquad \forall k \in K$$

Here, \(x_i\) is the decisions variable (e.g., specifying whether planning unit \(i\) has been selected (1) or not (0)), \(r_{ijk}\) is the amount of feature \(j\) in planning unit \(i\) in realization \(k\), \(t_j\) is the maximum representation target for feature \(j\) across all realizations \(k\), i.e., \(T_j = \max_k(t_{jk})\), \(y_j\) denotes the robust representation shortfall for the target \(t_j\) for feature \(j\) across all realizations \(k\), \(v_{jk}\) is the shortfall for feature \(j\) under realization \(k\), and \(w_j\) is the weight for feature \(j\) (defaults to 1 for all features; see prioritizr::add_feature_weights() to specify weights). \(\alpha\) is the specified conf_level confidence level for the uncertain constraint (as specified in add_*_robust_constraints), and ensures that the proportion of constraints for each feature group \(k\) that are held is higher than the specified confidence level. Additionally, \(B\) is the budget allocated for the solution, \(c_i\) is the cost of planning unit \(i\). Note that \(y_j\) is a continuous variable bounded between zero and one, and denotes the shortfall for target \(j\) as a proportion of the total target.

If conf_level = 1 (default), then the probabilistic constraint is simply represented as: $$ \sum_{i = 1}^{I} x_i r_{ijk} + T_{j} y_j \geq T_j \quad \forall \space j \in J, \space k \in K $$

If conf_level < 1, the probabilistic constraint is parameterised using a Chance Constraint Programming approach by reformulating the constraint into the following series of constraints: $$ \sum_{i = 1}^{I} x_i r_{ijk} + T_j y_j + T_j m_{jk} \geq T_{j} \quad \forall \space j \in J, \space k \in K \\ \sum_{k = 1}^{K_j} m_{jk} / K_j \leq 1 - \alpha \quad \forall \space j \in J\\ m \in \{0, 1\} $$

where \(m\) is a binary discrete variable indicating whether the constraint has been held or not, and \(K_j\) is the number of realizations for the feature \(j\). The auxiliary variable \(m\) is essentially a count of the number of times the probabilistic constraint has been violated (if \(m_{jk}=1\)), and the problem ensures that the total amount of times these are violated don't exceed \(1-\alpha\).

The "Conditional Value-at-Risk" method found in add_robust_min_set_objective has not yet been implemented in this objective function.

References

Charnes, A., & Cooper, W. W. (1959). Chance-Constrained Programming. Management Science, 6(1), 73-79.

See also

Examples

if (FALSE) { # \dontrun{
pu <- get_sim_pu_raster()

# Get feature data
features <- replicate(2, get_sim_features())
features <- rast(features)
names(features) <- paste0("feature_", rep(1:5, 2),
"_scenario_", rep(1:2, each = 5))
relative_budget <- as.numeric(global(pu, 'sum', na.rm = T)) * 0.1

# Get the groups
groups <- rep(paste0("feature_", 1:5), 2)

# Set up prioritizr problem
p <- problem(pu, features) %>%
  add_constant_robust_constraints(groups = groups) %>%
  add_robust_min_shortfall_objective(budget = relative_budget, method = "Chance") %>%
  add_relative_targets(0.1) %>%
  add_binary_decisions() %>%
  add_default_solver()

# Solve the problem
soln <- solve(p)

# Plot the solution
plot(soln)
} # }