Add variable robust constraints

Add robust constraints to a conservation problem to ensure that the priority areas are robust to uncertainty. In particular, this function is useful when the confidence level is different for different features. For example, this function may be especially useful when it is important to ensure that a prioritization is highly robust to uncertainty in the spatial distribution of threatened species, and only moderately robust to uncertainty in the spatial distribution of widespread species.

Usage

add_variable_robust_constraints(x, data)

Arguments

x

prioritizr::problem() object.

data

tibble::tibble() data frame containing information the feature groupings and their desired confidence level. See the Data format section for details. Also, see the Examples section for example usage.

Value

An updated prioritizr::problem() object with the constraint added to it.

Details

The robust/ chance constraints ensures that the proportion of constraints that are held is greater than a specified conf_level. If conf_level = 1, all constraints within the feature group must be held, meaning that the solution is fully robust to uncertainty. Lowering the conf_level to less thanuys 1 allows a certain percentage of the constraints for each feature group to be violated, enabling the algorithm to search of solutions with better objective values, while keeping the percentage of constraints violated less than 1 - conf_level.

Data format

The data argument must be a tibble::tibble() data frame that has information on the feature groupings and their confidence levels. Here, each row corresponds to a different feature group and columns contain information about the groups. In particular, it has the following columns.

features

A list column with the names of the features that belong to each group. In particular, if a particular set of features should belong to the same group, then they should be stored in the same element of this column.

conf_level

A numeric column with values that describe the confidence level associated with each feature group (ranging between 0 and 1). For example, a value of zero corresponds to a low confidence level, and so the optimization process will not constrained by any constraints in that particular group. Alternatively, a value of of one corresponds to a high level of robustness, and so the optimization process will be highly constrained to be robust against uncertainty for that particular group.

Data requirements

The robust constraints require that you have multiple alternative realizations for each biodiversity element of interest (e.g., species, ecosystems, ecosystem services). For example, we might have 5 species of interest. By applying different spatial modeling techniques, we might have 10 different models for each of the 5 different species. We can use these models to generate 10 alternative realizations for each of the 5 species (yielding 50 alternative realizations in total). To use these data, we would input these 50 alternative realizations as 50 features when initializing a conservation planning problem (i.e., prioritizr::problem()) and then use this function to specify which of the of the features correspond to the same species (based on the feature groupings parameter).

References

Charnes, A., & Cooper, W. W. (1959). Chance-constrained programming. Management Science, 6(1), 73–79.

See also

Examples

if (FALSE) { # \dontrun{
library(prioritizr)
library(terra)
library(tibble)

# Get planning unit data
pu <- get_sim_pu_raster()

# Get feature data
features <- replicate(2, get_sim_features())
features <- rast(features)
names(features) <- paste0("feature_", rep(1:5, 2), "_scenario_", rep(1:2, each = 5))
relative_budget <- as.numeric(global(pu, 'sum', na.rm = T)) * 0.1

# Get the groups
groups <- rep(paste0("feature_", 1:5), 2)

# Add variable robust constraints, specifying different confidence levels for each feature
variable_constraints <- tibble(
  features = split(names(features), groups),
  conf_level = c(1, 0.8, 0.6, 0.6, 0.5)
)

# Set up prioritizr problem
p <- problem(pu, features) %>%
  add_variable_robust_constraints(data = variable_constraints) %>%
  add_robust_min_set_objective() %>%
  add_relative_targets(0.1) %>%
  add_binary_decisions() %>%
  add_default_solver()

# Solve the problem
soln <- solve(p)

# Plot the solution
plot(soln)
} # }