robust.prioritizr: Robust Systematic Conservation Prioritization in R

The robust.prioritizr R package provides tools for building and solving robust systematic conservation prioritization problems. It extends the prioritizr package to account for uncertainty in the input data. This is particularly useful when working with data that is subject to change, such as species distribution models under climate change scenarios.

Details

This package contains several vignettes that are designed to showcase its functionality. To view them, please use the code vignette("name", package = "robust.prioritizr") where "name" is the name of the desired vignette (e.g., "robust.prioritizr").

robust.prioritizr

Brief introduction to the package.

climate-sdm

Example using simulated data from a species distribution model.

vic-cons-planning

Example using Victoria, Australia.

See also

Author

Authors:

Frankie Cho frankie.cho@qut.edu.au (ORCID)

Jeffrey O Hanson jeffrey.hanson@uqconnect.edu.au (ORCID)